+++ CPXnbExporter/XnbWriter.cs
@@ -371,7 +371,7 @@ namespace CPXnbExporter
 
                         if (n > 0)
                         {
-                            pixels[i] = new Color((byte)(r / n), (byte)(g / n), (byte)(b / n), 0);
+                            pixels[i] = new Color((byte)(r / n), (byte)(g / n), (byte)(b / n), (byte)0);
                             filled[i] = true;
                             changed = true;
                         }
